requirements.txt
scapy
sqlalchemy
rich
python-magic
dash-cytoscape